+++ src/ways.php
@@ -21,7 +21,7 @@ use RuntimeException;
  *
  * @package divengine/ways
  * @author  Rafa Rodriguez [@rafageist] <rafageist@hotmail.com>
- * @version 2.3.1
+ * @version 2.3.2
  *
  * @link    https://divengine.com
  * @link    https://divengine.com/ways
@@ -57,7 +57,7 @@ class ways
 
     const PROPERTY_RULES = 'rules';
 
-    private static $__version = '2.3.1';
+    private static $__version = '2.3.2';
 
     private static $__way_var;
 
@@ -1432,6 +1432,10 @@ class ways
                     }
                     foreach ($rules as $rule) {
                         if (!empty(self::$__controllers[$prop[self::PROPERTY_ID]])) {
+
+                            if (!isset(self::$__controllers[$prop[self::PROPERTY_ID]]['prop'][self::PROPERTY_RULES])){
+                                self::$__controllers[$prop[self::PROPERTY_ID]]['prop'][self::PROPERTY_RULES] = [];
+                            }
 
                             if (is_string(self::$__controllers[$prop[self::PROPERTY_ID]]['prop'][self::PROPERTY_RULES])) {
                                 self::$__controllers[$prop[self::PROPERTY_ID]]['prop'][self::PROPERTY_RULES] = [self::$__controllers[$prop[self::PROPERTY_ID]]['prop'][self::PROPERTY_RULES]];